+++ changelog.md
@@ -17,6 +17,7 @@
 
 ### Improvements
 
+- FLEX theme: pure CSS responsive tabs (tabs transform to accordion on screens smaller than md)
 - (DEV) New methods: TabFactory `InsertBefore()`,`InsertAfter()`, `InsertBeforeAny()`, `InsertAfterAny()`, `InsertAt()`
 - (DEV) New attribute for `tab` TagHelper: `sm-hide-if-empty`
 - (DEV) New rendering extension method: `IHtmlContent.HasValue()`